Improved rendering support for core and unembedded fonts
Core fonts are now better detected. The standard AFM metrics are used
for the glyph ax and ay attributes.

Separate sx and sy attributes have been added to glyphs. These represent
the advance x and y for the substituted font for core or other unembedded
fonts.

This gives more information for the renderer to make adjustments to font
size or scaling to better match the expected font positions.